created side menu component
