+++ components/SideMenu.vue
@@ -1,0 +1,105 @@
+<script setup>
+const isOpen = ref(false);
+</script>
+
+<template>
+  <USlideover v-model="isOpen" :overlay="false" side="left" width="5xl">
+    <div class="flex flex-col">
+      <ul>
+        <li class="h-6"></li>
+        <li>
+          <NuxtLink to="/">
+            <div
+              class="flex p-2 text-xl md:hover:bg-green-400 hover:text-black hover:cursor-pointer"
+            >
+              <svg
+                xmlns="http://www.w3.org/2000/svg"
+                viewBox="0 0 20 20"
+                fill="currentColor"
+                class="size-5 mt-1 mx-2"
+              >
+                <path
+                  fill-rule="evenodd"
+                  d="M9.293 2.293a1 1 0 0 1 1.414 0l7 7A1 1 0 0 1 17 11h-1v6a1 1 0 0 1-1 1h-2a1 1 0 0 1-1-1v-3a1 1 0 0 0-1-1H9a1 1 0 0 0-1 1v3a1 1 0 0 1-1 1H5a1 1 0 0 1-1-1v-6H3a1 1 0 0 1-.707-1.707l7-7Z"
+                  clip-rule="evenodd"
+                />
+              </svg>
+              <p>Home</p>
+            </div>
+          </NuxtLink>
+        </li>
+        <li>
+          <NuxtLink to="/create-campaign">
+            <div
+              class="flex p-2 text-xl hover:bg-green-400 hover:text-black hover:cursor-pointer"
+            >
+              <svg
+                xmlns="http://www.w3.org/2000/svg"
+                viewBox="0 0 20 20"
+                fill="currentColor"
+                class="size-5 mt-1 mx-2"
+              >
+                <path
+                  fill-rule="evenodd"
+                  d="M18 10a8 8 0 1 1-16 0 8 8 0 0 1 16 0Zm-1.5 0a6.5 6.5 0 1 1-11-4.69v.447a3.5 3.5 0 0 0 1.025 2.475L8.293 10 8 10.293a1 1 0 0 0 0 1.414l1.06 1.06a1.5 1.5 0 0 1 .44 1.061v.363a1 1 0 0 0 .553.894l.276.139a1 1 0 0 0 1.342-.448l1.454-2.908a1.5 1.5 0 0 0-.281-1.731l-.772-.772a1 1 0 0 0-1.023-.242l-.384.128a.5.5 0 0 1-.606-.25l-.296-.592a.481.481 0 0 1 .646-.646l.262.131a1 1 0 0 0 .447.106h.188a1 1 0 0 0 .949-1.316l-.068-.204a.5.5 0 0 1 .149-.538l1.44-1.234A6.492 6.492 0 0 1 16.5 10Z"
+                  clip-rule="evenodd"
+                />
+              </svg>
+              <p>Campaigns</p>
+            </div>
+          </NuxtLink>
+        </li>
+        <li>
+          <div
+            class="flex p-2 text-xl hover:bg-green-400 hover:text-black hover:cursor-pointer"
+          >
+            <svg
+              xmlns="http://www.w3.org/2000/svg"
+              viewBox="0 0 20 20"
+              fill="currentColor"
+              class="size-5 mt-1 mx-2"
+            >
+              <path
+                fill-rule="evenodd"
+                d="M4.5 2A1.5 1.5 0 0 0 3 3.5v13A1.5 1.5 0 0 0 4.5 18h11a1.5 1.5 0 0 0 1.5-1.5V7.621a1.5 1.5 0 0 0-.44-1.06l-4.12-4.122A1.5 1.5 0 0 0 11.378 2H4.5Zm2.25 8.5a.75.75 0 0 0 0 1.5h6.5a.75.75 0 0 0 0-1.5h-6.5Zm0 3a.75.75 0 0 0 0 1.5h6.5a.75.75 0 0 0 0-1.5h-6.5Z"
+                clip-rule="evenodd"
+              />
+            </svg>
+
+            <p>Characters</p>
+          </div>
+        </li>
+        <li>
+          <div
+            class="flex p-2 text-xl hover:bg-green-400 hover:text-black hover:cursor-pointer"
+          >
+            <svg
+              xmlns="http://www.w3.org/2000/svg"
+              viewBox="0 0 20 20"
+              fill="currentColor"
+              class="size-5 mt-1 mx-2"
+            >
+              <path
+                d="M10 8a3 3 0 1 0 0-6 3 3 0 0 0 0 6ZM3.465 14.493a1.23 1.23 0 0 0 .41 1.412A9.957 9.957 0 0 0 10 18c2.31 0 4.438-.784 6.131-2.1.43-.333.604-.903.408-1.41a7.002 7.002 0 0 0-13.074.003Z"
+              />
+            </svg>
+
+            <p>My Profile</p>
+          </div>
+        </li>
+      </ul>
+    </div>
+    <div class="p-4 flex-1">
+      <UButton
+        color="gray"
+        variant="ghost"
+        size="sm"
+        icon="i-heroicons-x-mark-20-solid"
+        class="flex sm:hidden absolute end-5 top-5 z-10"
+        square
+        padded
+        @click="isOpen = false"
+      />
+    </div>
+  </USlideover>
+</template>
